Introducing `query` that will let you query into the view hierarchy. Still very immature but quite powerful.
Can be combined with `touch` like this:

    touch(query("button"))

+++ ruby-gem/test-server/instrumentation-backend/src/sh/calaba/instrumentationbackend/InstrumentationBackend.java
@@ -1,5 +1,8 @@
 package sh.calaba.instrumentationbackend;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
+import com.jayway.android.robotium.solo.PublicViewFetcher;
 import sh.calaba.instrumentationbackend.actions.Actions;
 import sh.calaba.instrumentationbackend.actions.HttpServer;
 import android.app.Activity;
@@ -19,6 +22,7 @@ public class InstrumentationBackend extends ActivityInstrumentationTestCase2 {
     
     public static Instrumentation instrumentation;
     public static SoloEnhanced solo;
+    public static PublicViewFetcher viewFetcher;
     public static Actions actions;
 
     public InstrumentationBackend() {
@@ -30,6 +34,7 @@ public class InstrumentationBackend extends ActivityInstrumentationTestCase2 {
         super.setUp();
         
         solo = new SoloEnhanced(getInstrumentation(), this.getActivity());
+        viewFetcher = new PublicViewFetcher(getInstrumentation(), this.getActivity());
         actions = new Actions(getInstrumentation(), this);
         instrumentation = getInstrumentation();
     }
@@ -71,11 +76,14 @@ public class InstrumentationBackend extends ActivityInstrumentationTestCase2 {
         Log.e(TAG, message);
     }
 
-    private static void removeTestLocationProviders(Activity activity) {
-        final LocationManager locationService =
-            (LocationManager) activity.getSystemService(Context.LOCATION_SERVICE);
-        for (final String provider : locationService.getAllProviders()) {
-            locationService.removeTestProvider(provider);
+    private void removeTestLocationProviders(Activity activity) {
+        int hasPermission = solo.getCurrentActivity().checkCallingOrSelfPermission(Manifest.permission.ACCESS_MOCK_LOCATION);
+
+        if (hasPermission == PackageManager.PERMISSION_GRANTED) {
+            LocationManager locationService = (LocationManager) activity.getSystemService(Context.LOCATION_SERVICE);
+            for (final String provider : locationService.getAllProviders()) {
+                locationService.removeTestProvider(provider);
+            }
         }
     }
 }
